# api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenTunnel Control Plane API
  version: 1.0.0
  description: Control-plane contract shared by the server, CLI, and dashboard.
servers:
  - url: /api/v1
security:
  - cookieSession: []
  - bearerToken: []
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Error"
  /device/authorizations:
    post:
      operationId: createDeviceAuthorization
      security: []
      responses:
        "201":
          description: Device authorization created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorization"
        default:
          $ref: "#/components/responses/Error"
  /device/authorizations/{userCode}/approve:
    post:
      operationId: approveDeviceAuthorization
      parameters:
        - $ref: "#/components/parameters/UserCode"
      responses:
        "204":
          description: Device approved.
        default:
          $ref: "#/components/responses/Error"
  /device/token:
    post:
      operationId: exchangeDeviceCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceTokenRequest"
      responses:
        "200":
          description: CLI token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceToken"
        "202":
          description: Authorization remains pending.
          headers:
            Retry-After:
              schema:
                type: integer
        default:
          $ref: "#/components/responses/Error"
  /domains:
    get:
      operationId: listDomains
      responses:
        "200":
          description: Owned domains.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createDomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainRequest"
      responses:
        "201":
          description: Domain allocated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        default:
          $ref: "#/components/responses/Error"
  /domains/{domainId}:
    get:
      operationId: getDomain
      parameters:
        - $ref: "#/components/parameters/DomainId"
      responses:
        "200":
          description: Domain details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteDomain
      parameters:
        - $ref: "#/components/parameters/DomainId"
      responses:
        "204":
          description: Domain released.
        default:
          $ref: "#/components/responses/Error"
...
  /tokens:
    get:
      operationId: listTokens
      responses:
        "200":
          description: Revocable CLI tokens.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenSummary"
        default:
          $ref: "#/components/responses/Error"
  /tokens/{tokenId}:
    delete:
      operationId: revokeToken
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Token revoked.
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: opentunnel_session
    bearerToken:
      type: http
      scheme: bearer
  parameters:
    UserCode:
      name: userCode
      in: path
      required: true
      schema:
        type: string
    DomainId:
      name: domainId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
...
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    DeviceAuthorization:
      type: object
      required:
        [
          deviceCode,
          userCode,
          verificationUri,
          verificationUriComplete,
          expiresIn,
          interval,
        ]
      properties:
        deviceCode:
          type: string
        userCode:
          type: string
        verificationUri:
          type: string
          format: uri
        verificationUriComplete:
          type: string
          format: uri
        expiresIn:
          type: integer
        interval:
          type: integer
    DeviceTokenRequest:
      type: object
      required: [deviceCode]
      properties:
        deviceCode:
          type: string
    DeviceToken:
      type: object
      required: [accessToken, tokenType]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
    CreateDomainRequest:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
    Domain:
      type: object
      required: [id, slug, hostname, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        hostname:
          type: string
        status:
          type: string
          enum: [offline, online]
        createdAt:
          type: string
          format: date-time
...
    TokenSummary:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        lastUsedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time